tests/unit/test_processors/test_formula_processor.py
# tests/unit/test_processors/test_formula_processor.py
import pytest
import pandas as pd
import numpy as np
from excel_processor.processors.formula_processor import FormulaProcessor
from excel_processor.models.formula import Formula, FormulaType

@pytest.fixture
def sample_config():
    return {
        'test': 'config'
    }

@pytest.fixture
def sample_data():
    return {
        'Sheet1': pd.DataFrame({
            'A': [1, 2, 3],
            'B': [4, 5, 6],
            'C': [None, None, None]
        })
    }

def test_formula_processor_initialization(sample_config):
    """Test FormulaProcessor initialization."""
    processor = FormulaProcessor(sample_config)
    assert processor.config == sample_config
    assert isinstance(processor.formula_cache, dict)
    assert len(processor.formula_cache) == 0

def test_process_empty_formula_cache(sample_config, sample_data):
    """Test processing with empty formula cache."""
    processor = FormulaProcessor(sample_config)
    result = processor.process(sample_data)
    assert 'Sheet1' in result
    pd.testing.assert_frame_equal(result['Sheet1'], sample_data['Sheet1'])

def test_process_with_arithmetic_formula(sample_config, sample_data):
    """Test processing with arithmetic formula."""
    processor = FormulaProcessor(sample_config)
    
    # Add formula to cache
    formula = Formula(
        raw_formula='=A + B',
        python_equivalent='df["A"] + df["B"]',
        formula_type=FormulaType.ARITHMETIC,
        dependencies={'A', 'B'},
        sheet_name='Sheet1',
        column_name='C'
    )
    processor.formula_cache['Sheet1.C'] = formula
    
    result = processor.process(sample_data)
    expected = [5, 7, 9]  # A + B
    np.testing.assert_array_equal(result['Sheet1']['C'], expected)

def test_formula_evaluation_error(sample_config, sample_data):
    """Test handling of formula evaluation errors."""
    processor = FormulaProcessor(sample_config)
    
    # Add invalid formula to cache
    formula = Formula(
        raw_formula='=Invalid',
        python_equivalent='invalid_code',
        formula_type=FormulaType.ARITHMETIC,
        dependencies=set(),
        sheet_name='Sheet1',
        column_name='C'
    )
    processor.formula_cache['Sheet1.C'] = formula
    
    with pytest.raises(ValueError, match="Formula evaluation error"):
        processor.process(sample_data)

def test_lookup_formula_handling(sample_config):
    """Test handling of lookup formulas."""
    processor = FormulaProcessor(sample_config)
    
    formula = Formula(
        raw_formula='=VLOOKUP(A1, Sheet2!A:B, 2, FALSE)',
        python_equivalent='# lookup implementation',
        formula_type=FormulaType.LOOKUP,
        dependencies={'A1', 'Sheet2!A:B'},
        sheet_name='Sheet1',
        column_name='Result'
    )
    
    context = {
        'np': np,
        'pd': pd,
        'current': pd.DataFrame(),
        'data': {}
    }
    
    # Since _handle_lookup_formula is not implemented, it should pass
    result = processor._handle_lookup_formula(formula, context)
    assert result is None

def test_aggregate_formula_handling(sample_config):
    """Test handling of aggregate formulas."""
    processor = FormulaProcessor(sample_config)
    
    formula = Formula(
        raw_formula='=SUM(A1:A10)',
        python_equivalent='np.sum(df["A"])',
        formula_type=FormulaType.AGGREGATE,
        dependencies={'A1:A10'},
        sheet_name='Sheet1',
        column_name='Result'
    )
    
    context = {
        'np': np,
        'pd': pd,
        'current': pd.DataFrame(),
        'data': {}
    }
    
    # Since _handle_aggregate_formula is not implemented, it should pass
    result = processor._handle_aggregate_formula(formula, context)
    assert result is None